create fire indices submodule, rename fire_weather_indexes to cffwis_indices, add appropriate DeprecationWarnings for indice and indicator, move ffdi to fire submodule, rename fwi to cffwis and place in fire submodule, adjust tests to account for new calls and paths
