+++ xclim/testing/tests/test_indicators.py
@@ -476,11 +476,11 @@ def test_doc():
     doc = xclim.atmos.fire_weather_indexes.__doc__
     assert doc.startswith("Fire weather indexes. (realm: atmos)")
     assert "This indicator will check for missing values according to the method" in doc
-    assert "Based on indice :py:func:`~xclim.indices.fwi.fire_weather_indexes`." in doc
+    assert "Based on indice :py:func:`~xclim.indices.fire.fire_weather_indexes`." in doc
     assert "ffmc0 : str or DataArray, optional" in doc
     assert "Returns\n-------" in doc
     assert "See :cite:t:`code-natural_resources_canada_data_nodate`, " in doc
-    assert "the :py:mod:`xclim.indices.fwi` module documentation," in doc
+    assert "the :py:mod:`xclim.indices.fire` module documentation," in doc
     assert (
         "and the docstring of :py:func:`fire_weather_ufunc` for more information."
         in doc
